Back/Ver-1.0/web/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django
django-crontab==0.7.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3>=1.26.5
